1. 修复类型不兼容的 Bug，导致数据读取时候会因为类型长度不足无法读取。

+++ dbvisitor-faker/src/main/java/net/hasor/dbvisitor/faker/generator/provider/carefully/PostgresCarefullyTypeSrwFactory.java
@@ -93,7 +93,7 @@ public class PostgresCarefullyTypeSrwFactory extends AbstractPostgresTypeSrwFact
             case "int8": {
                 NumberSeedFactory seedFactory = new NumberSeedFactory();
                 NumberSeedConfig seedConfig = seedFactory.newConfig();
-                seedConfig.setNumberType(NumberType.Integer);
+                seedConfig.setNumberType(NumberType.Long);
                 seedConfig.addMinMax(new BigDecimal("0"), new BigDecimal("100000000"));
                 return finalSrw(seedFactory, seedConfig, Types.BIGINT, isArray, columnConfig, columnType);
             }
